+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/assets/config/franka.py
@@ -7,7 +7,8 @@
 
 The following configurations are available:
 
-* :obj:`FRANKA_PANDA_ARM_WITH_PANDA_HAND_CFG`: Franka Emika Panda robot with Panda hand
+* :obj:`FRANKA_PANDA_CFG`: Franka Emika Panda robot with Panda hand
+* :obj:`FRANKA_PANDA_HIGH_PD_CFG`: Franka Emika Panda robot with Panda hand with stiffer PD control
 
 Reference: https://github.com/frankaemika/franka_ros
 """
@@ -22,7 +23,7 @@ from ..articulation import ArticulationCfg
 # Configuration
 ##
 
-FRANKA_PANDA_ARM_WITH_PANDA_HAND_CFG = ArticulationCfg(
+FRANKA_PANDA_CFG = ArticulationCfg(
     spawn=sim_utils.UsdFileCfg(
         usd_path=f"{ISAAC_ORBIT_NUCLEUS_DIR}/Robots/FrankaEmika/panda_instanceable.usd",
         activate_contact_sensors=False,
@@ -51,14 +52,14 @@ FRANKA_PANDA_ARM_WITH_PANDA_HAND_CFG = ArticulationCfg(
         "panda_shoulder": ImplicitActuatorCfg(
             joint_names_expr=["panda_joint[1-4]"],
             effort_limit=87.0,
-            velocity_limit=100.0,
+            velocity_limit=2.175,
             stiffness=80.0,
             damping=4.0,
         ),
         "panda_forearm": ImplicitActuatorCfg(
             joint_names_expr=["panda_joint[5-7]"],
-            effort_limit=87.0,
-            velocity_limit=100.0,
+            effort_limit=12.0,
+            velocity_limit=2.61,
             stiffness=80.0,
             damping=4.0,
         ),
@@ -72,4 +73,16 @@ FRANKA_PANDA_ARM_WITH_PANDA_HAND_CFG = ArticulationCfg(
     },
     soft_joint_pos_limit_factor=1.0,
 )
-"""Configuration of Franka Emika Panda robot with Panda hand."""
+"""Configuration of Franka Emika Panda robot."""
+
+
+FRANKA_PANDA_HIGH_PD_CFG = FRANKA_PANDA_CFG.copy()
+FRANKA_PANDA_HIGH_PD_CFG.spawn.rigid_props.disable_gravity = True
+FRANKA_PANDA_HIGH_PD_CFG.actuators["panda_shoulder"].stiffness = 400.0
+FRANKA_PANDA_HIGH_PD_CFG.actuators["panda_shoulder"].damping = 80.0
+FRANKA_PANDA_HIGH_PD_CFG.actuators["panda_forearm"].stiffness = 400.0
+FRANKA_PANDA_HIGH_PD_CFG.actuators["panda_forearm"].damping = 80.0
+"""Configuration of Franka Emika Panda robot with stiffer PD control.
+
+This configuration is useful for task-space control using differential IK.
+"""