+++ source/extensions/omni.isaac.orbit/test/assets/test_articulation.py
@@ -30,7 +30,7 @@ import omni.isaac.orbit.sim as sim_utils
 import omni.isaac.orbit.utils.string as string_utils
 from omni.isaac.orbit.actuators import ImplicitActuatorCfg
 from omni.isaac.orbit.assets import Articulation, ArticulationCfg
-from omni.isaac.orbit.assets.config import ANYMAL_C_CFG, FRANKA_PANDA_ARM_WITH_PANDA_HAND_CFG
+from omni.isaac.orbit.assets.config import ANYMAL_C_CFG, FRANKA_PANDA_CFG
 from omni.isaac.orbit.utils.assets import ISAAC_NUCLEUS_DIR
 
 
@@ -120,7 +120,7 @@ class TestArticulation(unittest.TestCase):
     def test_initialization_fixed_base(self):
         """Test initialization for fixed base."""
         # Create articulation
-        robot = Articulation(cfg=FRANKA_PANDA_ARM_WITH_PANDA_HAND_CFG.replace(prim_path="/World/Robot"))
+        robot = Articulation(cfg=FRANKA_PANDA_CFG.replace(prim_path="/World/Robot"))
 
         # Check that boundedness of articulation is correct
         self.assertEqual(ctypes.c_long.from_address(id(robot)).value, 1)